LayoutTests/fast/xmlhttprequest/xmlhttprequest-responseXML-html-no-responsetype.html

<!DOCTYPE html>
<html>
<head>
    <script src="../js/resources/js-test-pre.js"></script>
    <script>
        description('This tests the XMLHttpRequest responseXML loading an HTML document with no specified responseType.');

        window.jsTestIsAsync = true;
        var xhr = new XMLHttpRequest(),
            url = window.location.href;
        xhr.onload = function() {
            shouldBeNull('xhr.responseXML');
            finishJSTest();
        };
        xhr.onerror = function() {
            testFailed('The XHR request to an existing resource failed: "' + url + '"');
            finishJSTest();
        };
        xhr.open('GET', url);
        xhr.send(null);
    </script>
    <script src="../js/resources/js-test-post.js"></script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
